Only include logging.h in cc files.
